chore(page-slice): Rename shit to PageSection
