+++ src/components/PageSection/svg/CurvedSvg.tsx
