+++ src/views/TradingCompetition/pageSectionStyles.tsx
